Database

+++ Models/CreateMenuItemDTO.cs
@@ -4,5 +4,5 @@ public class CreateMenuItemDTO
 {
     public Guid Id { get; set; }
     public string Name { get; set; }
-    public int? Price { get; set; }
+    public Double? Price { get; set; }
 }
